third_party/blink/web_tests/inspector-protocol/runtime/runtime-evaluate-return-by-value-expected.txt

Tests that Runtime.evaluate returns an error for non serialable types.

Running test: testSymbol
{
    error : {
        code : -32000
        message : Object couldn't be returned by value
    }
    id : <number>
    sessionId : <string>
}

Running test: testObjectWithChain
{
    error : {
        code : -32000
        message : Object reference chain is too long
    }
    id : <number>
    sessionId : <string>
}
